# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Conflux Web Monitoring API
  description: REST API for monitoring and controlling Conflux orchestration
  contact:
    name: Conflux
    url: https://github.com/anomalyco/conflux
  license:
    name: MIT
    identifier: MIT
  version: 0.6.205
paths:
...
  /api/changes/{id}:
    get:
      tags:
      - changes
      summary: Get a specific change by ID
      operationId: get_change
      parameters:
      - name: id
        in: path
        description: Change ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Change found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeStatus'
        '404':
          description: Change not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v2/changes/{change_id}:
    get:
      tags:
      - remote-control
      summary: One projected change.
      operationId: get_change
      parameters:
      - name: change_id
        in: path
        description: Change ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Projected change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeResponse'
        '404':
          description: No such change in this incarnation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v2/commands:
    post:
      tags:
      - remote-control
      summary: Submit a command from the closed v2 command set.
      operationId: submit_command
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandRequest'
        required: true
      responses:
        '200':
          description: Completed synchronously, replayed, or an explicit no-op
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandRecord'
        '202':
          description: Accepted and still running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandRecord'
        '409':
          description: Stale revision, lifecycle, eligibility, or idempotency conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Schema or correlation-ID validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: No record slot could be reserved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v2/commands/{command_id}:
    get:
      tags:
      - remote-control
      summary: Look up a previously submitted command.
      operationId: get_command
      parameters:
      - name: command_id
        in: path
        description: Command ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Command record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandRecord'
        '404':
          description: No such command in this incarnation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      description: Structured error body returned by every non-success v2 response.
      required:
      - error_code
      - message
      - correlation_id
      properties:
        correlation_id:
          type: string
          description: Correlation ID of the failing request (caller-supplied or generated).
        current_revision:
          type:
          - integer
          - 'null'
          format: int64
          description: Current state revision, when the failure is revision-related.
          minimum: 0
        error_code:
          $ref: '#/components/schemas/ErrorCode'
          description: Stable code the client branches on.
        message:
          type: string
          description: Sanitized human-readable detail.
...
    ChangeResource:
      type: object
      description: One change as projected into the v2 snapshot.
      required:
      - id
      - display_status
      - progress_status
      - completed_tasks
      - total_tasks
      - progress_percent
      - dependencies
      properties:
        completed_tasks:
          type: integer
          format: int32
          description: Completed tasks.
          minimum: 0
        dependencies:
          type: array
          items:
            type: string
          description: Declared dependencies.
        display_status:
          type: string
          description: Reducer-derived display status (canonical taxonomy).
        id:
          type: string
          description: Change ID.
        iteration_number:
          type:
          - integer
          - 'null'
          format: int32
          description: Apply/archive iteration number, when a loop is running.
          minimum: 0
        progress_percent:
          type: number
          format: float
          description: Progress percentage (0-100).
        progress_status:
          type: string
          description: Task-progress status (`pending`, `in_progress`, `complete`).
        total_tasks:
          type: integer
          format: int32
          description: Total tasks.
          minimum: 0
    ChangeResponse:
      type: object
      description: '`GET /api/v2/changes/{change_id}` body.'
      required:
      - instance_id
      - state_revision
      - change
      properties:
        change:
          $ref: '#/components/schemas/ChangeResource'
          description: The projected change.
        instance_id:
          type: string
          description: Process incarnation ID.
        state_revision:
          type: integer
          format: int64
          description: Revision the change was read at.
          minimum: 0
    ChangeStatus:
      type: object
      description: Change status for WebSocket updates
      required:
      - id
      - completed_tasks
      - total_tasks
      - progress_percent
      - status
      - dependencies
      properties:
        completed_tasks:
          type: integer
          format: int32
          description: Number of completed tasks
          minimum: 0
        dependencies:
          type: array
          items:
            type: string
          description: Dependencies on other changes
        id:
          type: string
          description: Change ID
        iteration_number:
          type:
          - integer
          - 'null'
          format: int32
          description: Current iteration number for apply/archive loops
          minimum: 0
        progress_percent:
          type: number
          format: float
          description: Progress percentage (0-100)
        queue_status:
          type:
          - string
          - 'null'
          description: |-
            Queue status (for parallel/serial execution tracking)
            Aligned with canonical display taxonomy values: "not queued", "queued", "blocked", "stalled", "applying",
            "accepting", "archiving", "archived", "merged", "pushed", "rejected", "merge wait", "resolving", "resolve pending", "reject pending", "error"
        status:
          type: string
          description: 'Current status: "pending", "in_progress", "complete"'
        total_tasks:
          type: integer
          format: int32
          description: Total number of tasks
          minimum: 0
...
    CommandRecord:
      type: object
      description: Command record returned by `POST /api/v2/commands` and by status lookup.
      required:
      - command_id
      - instance_id
      - type
      - state
      - expected_revision
      - correlation_id
      - idempotency_key
      - created_at
      properties:
        command_id:
          type: string
          description: Random 128-bit hexadecimal command ID.
        completed_at:
          type:
          - string
          - 'null'
          description: Completion time (RFC 3339), when finished.
        correlation_id:
          type: string
          description: Correlation label of the submitting request.
        created_at:
          type: string
          description: Submission time (RFC 3339).
        detail:
          type:
          - string
          - 'null'
          description: Sanitized outcome detail.
        error_code:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ErrorCode'
            description: Typed failure code when `state` is `failed`.
        expected_revision:
          type: integer
          format: int64
          description: Revision the command was admitted against.
          minimum: 0
        idempotency_key:
          type: string
          description: Replay key this record is bound to.
        instance_id:
          type: string
          description: Process incarnation that owns this record.
        result_revision:
          type:
          - integer
          - 'null'
          format: int64
          description: Revision observed once the command finished.
          minimum: 0
        state:
          $ref: '#/components/schemas/CommandState'
          description: Current lifecycle state.
        type:
          type: string
          description: Wire discriminant of the submitted command.
    CommandRequest:
      allOf:
      - $ref: '#/components/schemas/CommandSpec'
        description: Discriminated command payload.
      - type: object
        required:
        - expected_revision
        - idempotency_key
        properties:
          correlation_id:
            type:
            - string
            - 'null'
            description: Optional opaque trace label.
          expected_revision:
            type: integer
            format: int64
            description: Revision the caller believes is current.
            minimum: 0
          idempotency_key:
            type: string
            description: Caller-chosen replay key, unique per intended side effect.
      description: |-
        `POST /api/v2/commands` request envelope.

        Every command is side-effecting, so `expected_revision` and
        `idempotency_key` are mandatory for all of them without exception.
...
    CommandState:
      type: string
      description: Lifecycle of a submitted command within this process incarnation.
      enum:
      - running
      - succeeded
      - no_op
      - failed
...
    ErrorCode:
      type: string
      description: |-
        Stable, machine-actionable v2 error codes.

        The HTTP status classifies the transport failure; the client branches on
        `error_code`. Two different causes may share status 409, so the code — not
        the status — is the contract.
      enum:
      - unauthorized
      - forbidden
      - not_found
      - stale_revision
      - lifecycle_conflict
      - target_ineligible
      - root_busy
      - idempotency_mismatch
      - registry_capacity
      - validation_failed
      - internal_error
    ErrorResponse:
      type: object
      description: Error response for API errors
      required:
      - error
      properties:
        error:
          type: string
...